dev_requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=- --strip-extras dev_requirements.in > dev_requirements.txt
#
attrs==25.3.0
    # via
    #   outcome
    #   trio
blinker==1.9.0
    # via flask
build==1.2.2.post1
    # via pip-tools
certifi==2025.4.26
    # via
    #   requests
    #   selenium
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   flask
    #   pip-tools
flask==3.1.0
    # via
    #   -r dev_requirements.in
    #   flask-cors
flask-cors==5.0.1
    # via -r dev_requirements.in
h11==0.16.0
    # via wsproto
idna==3.10
    # via
    #   requests
    #   trio
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==25.0
    # via
    #   build
    #   pytest
    #   webdriver-manager
pip-tools==7.4.1
    # via -r dev_requirements.in
pluggy==1.5.0
    # via pytest
psutil==7.0.0
    # via -r dev_requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pysocks==1.7.1
    # via urllib3
pytest==8.3.5
    # via
    #   -r dev_requirements.in
    #   pytest-mock
pytest-mock==3.14.0
    # via -r dev_requirements.in
python-dotenv==1.1.0
    # via
    #   -r dev_requirements.in
    #   webdriver-manager
requests==2.32.3
    # via webdriver-manager
selenium==4.31.0
    # via -r dev_requirements.in
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.30.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typing-extensions==4.13.2
    # via selenium
uptime==3.0.1
    # via -r dev_requirements.in
urllib3==2.4.0
    # via
    #   requests
    #   selenium
webdriver-manager==4.0.2
    # via -r dev_requirements.in
websocket-client==1.8.0
    # via selenium
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
wheel==0.45.1
    # via pip-tools
wsproto==1.2.0
    # via trio-websocket

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
